refactor(mcp): move capability-gating helpers to utils.rs / utils_test.rs
Per @peicodes' review nit on #9436, extract `should_query_resources`
and `should_query_tools` from `native.rs` into a sibling `utils` module
following the codebase's `mod_test`-style convention (helper module +
external test file declared from the parent).

No behavior change; the spawn_server callsites continue to use the
predicates via `super::utils::*` and the regression test
`each_capability_is_queried_independently` is preserved verbatim in the
new test file.
